gpu optimized

+++ agent_dqn/model.py
@@ -6,7 +6,7 @@ def build_q_network(input_dim, output_dim):
         layers.Input(shape=(input_dim,)),
         layers.Dense(64, activation='relu'),
         layers.Dense(64, activation='relu'),
-        layers.Dense(output_dim, dtype='float32') # Why is there no (linear) activation in the output vector?
+        layers.Dense(output_dim) # Why is there no (linear) activation in the output vector?
